#860 widen pkt_meta.pkt_len to u32; fix IPv6 wrap (#893)
* The SCREEN_PING_OF_DEATH check was dead code:

  if (meta->pkt_len > 65535) ...

pkt_len was __u16, so the comparison was always false and
GLOBAL_CTR_SCREEN_PING_OF_DEATH never incremented. Same dead code
in xdp_screen.c and tc_screen_egress.c.

Widen to __u32. With u32, the check is now meaningful — fires for
oversized frames (jumbo / GSO segmented) that legitimately exceed
65535 bytes. The classical ping-of-death attack requires fragment
reassembly to overflow 65535; xpf doesn't reassemble, so this is
not a complete reimplementation, but the comparison is no longer
trivially-false dead code.

Also fixes the IPv6 wrap noted in the issue:

  meta->pkt_len = bpf_ntohs(ip6h->payload_len) + 40;
  meta->pkt_len = bpf_ntohs(ip6h->payload_len) + sizeof(struct ipv6hdr);

bpf_ntohs returns u16; the addition wrapped u16 even when the
target was u32. Cast to u32 first.

Closes #860.


* #860 SCREEN_PING_OF_DEATH: detect via IPv4 fragment offset+payload

Codex round 1 caught: my earlier u16→u32 widening for pkt_len did
not actually make the SCREEN_PING_OF_DEATH check functional. The
check was \`pkt_len > 65535\`, and pkt_len is derived from IP wire
fields (\`tot_len\` for IPv4, \`payload_len + 40\` for IPv6) — both
capped at u16 by the wire format. xpf doesn't reassemble
fragments, so a single received packet's pkt_len can never
exceed 65535. The widening fixed "always-false-due-to-type" but
left "always-false-due-to-bounded-input."

Real fix: detect the classical ping-of-death attack by checking
the per-fragment offset+payload against the 65535-byte reassembled
datagram limit. If \`(frag_offset_bytes + frag_payload) > 65535\`,
any reassembling host would reject — drop the fragment now.

Implemented in both xdp_screen.c and tc_screen_egress.c with
identical logic. IPv4 only; IPv6 ping-of-death detection requires
parsing the IPv6 fragment header (NEXTHDR_FRAGMENT) and is filed
as a follow-up.

The pkt_len type widening + IPv6 wrap fix from the prior commit
remain — they're standalone correctness improvements (the
checksum helper at xpf_helpers.h:436/598 still uses u16 l4_len
which fits for normal traffic; jumbo IPv6 would need widening
there too, also follow-up).


* #860 PoD: include IP header in overflow threshold (offset + tot_len)

Codex round 2 caught the off-by-header miss: my prior formula
checked `offset_bytes + frag_payload > 65535` which strips the IP
header from frag_payload. The reassembled IP datagram total =
ihl_bytes + max(offset + frag_payload). For overflow:
  offset_bytes + frag_payload + ihl_bytes > 65535
which simplifies to:
  offset_bytes + tot_len > 65535

(since tot_len = ihl_bytes + frag_payload). Drop the
intermediate ihl_bytes / frag_payload computation, just compare
offset_bytes + tot_len directly.

An attacker could have crafted the final fragment so
`offset + frag_payload` was just under 65535 while the actual
reassembled total (with the 20+ byte IP header) overflowed.


* #860 PoD: use worst-case first-IHL=60 threshold for cross-fragment safety

Codex round 3: per RFC 791 non-first fragments may omit non-copied
IP options, so their ihl can be smaller than the first fragment's.
The previous formula `offset + tot_len > 65535` trusts THIS
fragment's ihl (embedded in tot_len) to estimate the IP header
overhead — but that undercounts when the first fragment has
options.

Example: first fragment IHL=60, non-first at offset=65480 with
ihl=20 and tot_len=21 (1 byte payload). Round 2 check:
65480+21=65501 ≤ 65535, no drop. But reassembled total =
60+65481 = 65541, overflow.

Fix: use worst-case first IHL = 60 (RFC max). Threshold:
`offset + frag_payload > 65535 - 60 = 65475`. frag_payload uses
THIS fragment's actual ihl. Catches all reassembly overflows
regardless of options layout. False-positives only on legal
max-length datagrams that use 40+ bytes of IP options —
essentially never seen in modern traffic.

Same fix on TC egress path; comment in tc_screen_egress.c
points to xdp_screen.c for the full rationale.

Round 1 widened pkt_len; round 2 fixed off-by-header; round 3
fixes IHL-mismatch.


* #860 PoD: revert to threshold 65535 (no false positives on legal traffic)

Codex round 4: my round-3 worst-case-IHL=60 threshold (65475)
false-positives on legitimate near-max IPv4 datagrams. Concrete
example: 44×1480-byte fragments → last fragment offset=65120,
payload=395, sum=65515 > 65475 → DROP. That's a real legal
datagram, not a corner case.

Trade-off:
- Threshold 65475: catches cross-IHL exploit, breaks legal
  near-max traffic.
- Threshold 65535 with this fragment's tot_len: zero false
  positives, misses cross-IHL exploit.

For an OPERATOR-ENABLED screen, false-positives on legitimate
traffic are MUCH worse than missing exotic attacks. Operators
expect screens to drop attacks, not legitimate near-max IP
datagrams.

Mitigation for the cross-IHL gap: SCREEN_IP_SOURCE_ROUTE
already drops any IP packet with ihl>5. Operators concerned
about cross-IHL exploitation should enable BOTH screens. IP
options are also blocked by most middleboxes today.

Settled formula: `offset_bytes + tot_len > 65535` (round 2).
Comment block now spells out the trade-off so future readers
don't re-litigate it.


* #860 trim noisy comment per Codex round 5

Codex round 5: trade-off explanation is valuable but the
'round-N iteration' process diary belongs in commit log, not
production C comment. Trimmed to keep technical rationale +
SCREEN_IP_SOURCE_ROUTE mitigation + IPv6-follow-up note.


---------

+++ bpf/headers/xpf_helpers.h
@@ -292,7 +292,10 @@ parse_ipv6hdr(void *data, void *data_end, struct pkt_meta *meta)
 
 	meta->ip_ttl      = ip6h->hop_limit;
 	meta->dscp        = (ip6h->priority << 2) | (ip6h->flow_lbl[0] >> 6);
-	meta->pkt_len     = bpf_ntohs(ip6h->payload_len) + 40;
+	/* #860: cast to u32 before the +40 — bpf_ntohs returns u16; the
+	 * sum can wrap u16 for payload_len > 65495 even though pkt_len
+	 * is now u32. */
+	meta->pkt_len     = (__u32)bpf_ntohs(ip6h->payload_len) + 40;
 	meta->addr_family = AF_INET6;
 	meta->is_fragment = 0;
 
@@ -534,7 +537,8 @@ parse_ipv6_l4_fast(void *data, void *data_end, struct pkt_meta *meta)
 	meta->protocol = nexthdr;
 	meta->ip_ttl = ip6h->hop_limit;
 	meta->dscp = (ip6h->priority << 2) | (ip6h->flow_lbl[0] >> 6);
-	meta->pkt_len = bpf_ntohs(ip6h->payload_len) + sizeof(struct ipv6hdr);
+	/* #860: cast to u32 to avoid u16 wrap for jumbo frames. */
+	meta->pkt_len = (__u32)bpf_ntohs(ip6h->payload_len) + sizeof(struct ipv6hdr);
 	meta->addr_family = AF_INET6;
 	meta->is_fragment = 0;
 	meta->l4_offset = meta->l3_offset + sizeof(struct ipv6hdr);
